ECommerce/requirements.txt
aiokafka==0.5.2
amqp==2.5.2
APScheduler==3.6.3
asgiref==3.2.3
attrs==19.3.0
billiard==3.6.2.0
boto3==1.11.2
botocore==1.14.2
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
CodeConvert==3.0.2
colorama==0.4.3
ConfigArgParse==1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.2
Django==3.0.2
django-admin==2.0.0
django-cors-headers==3.2.0
django-elasticsearch-dsl==7.1.1
django-excel-response2==2.0.8
django-models-ext==1.1.9
django-nose==1.4.6
django-otp==0.8.1
django-redis==4.11.0
django-rest-swagger==2.2.0
django-short-url==1.1.4
django-six==1.0.4
django-storages==1.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
Faker==0.9.1
Flask==1.1.1
fun==1.0.0
furl==2.1.0
gevent==1.5a2
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
httplib2==0.9.2
idna==2.8
importlib-metadata==1.3.0
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
kafka-python==1.4.6
kombu==4.6.7
locustio==0.14.4
MarkupSafe==1.1.1
mixer==6.1.3
mock==3.0.5
more-itertools==8.0.2
msgpack-python==0.5.6
nose==1.3.7
openapi-codec==1.3.2
orderedmultidict==1.0.1
packaging==20.0
Pillow==7.0.0
pinocchio==0.4.2
pluggy==0.13.1
psutil==5.6.7
psycopg2==2.8.4
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.2
pytest-cov==2.8.1
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==3.7.0
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
pyzmq==18.1.1
redis==3.3.11
requests==2.22.0
s3transfer==0.3.0
screen==1.0.1
short-url==1.2.2
shortuuid==0.5.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.2
TimeConvert==1.5.1
twilio==6.38.0
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.8
Werkzeug==1.0.0
xlwt==1.3.0
zipp==0.6.0
